move tests into test/src directory
