+++ v-next/hardhat/test/fixture-projects/run-js-script/scripts/success.js
